+++ crates/cxx-qt-gen/src/generator/cpp/constructor.rs
@@ -17,8 +17,8 @@ fn default_constructor(
     base_class: String,
     initializers: String,
 ) -> GeneratedCppQObjectBlocks {
-    GeneratedCppQObjectBlocks {
-        methods: vec![CppFragment::Pair {
+    let constructor = if qobject.has_qobject_macro {
+        CppFragment::Pair {
             header: format!(
                 "explicit {class_name}(QObject* parent = nullptr);",
                 class_name = qobject.ident
@@ -34,7 +34,33 @@ fn default_constructor(
                 namespace_internals = qobject.namespace_internals,
                 rust_obj = qobject.rust_ident,
             ),
-        }],
+        }
+    } else {
+        CppFragment::Pair {
+            header: format!("explicit {class_name}();", class_name = qobject.ident),
+            source: formatdoc!(
+                r#"
+            {class_name}::{class_name}()
+              {base_class_line}
+              , ::rust::cxxqt1::CxxQtType<{rust_obj}>(::{namespace_internals}::createRs()){initializers}
+            {{ }}
+            "#,
+                base_class_line = if base_class.is_empty() {
+                    unreachable!(
+                        "Cannot have an empty #[base] attribute  with no #[qobject] attribute"
+                    );
+                } else {
+                    format!(": {base_class}()")
+                },
+                class_name = qobject.ident,
+                namespace_internals = qobject.namespace_internals,
+                rust_obj = qobject.rust_ident,
+            ),
+        }
+    };
+
+    GeneratedCppQObjectBlocks {
+        methods: vec![constructor],
         ..Default::default()
     }
 }
@@ -136,6 +162,12 @@ mod tests {
 
     use syn::parse_quote;
 
+    fn type_names_with_qobject() -> TypeNames {
+        let mut type_names = TypeNames::mock();
+        type_names.insert("QObject", None, None, None);
+        type_names
+    }
+
     fn qobject_for_testing() -> GeneratedCppQObject {
         GeneratedCppQObject {
             ident: "MyObject".to_string(),
@@ -143,6 +175,7 @@ mod tests {
             namespace: "".to_string(),
             namespace_internals: "rust".to_string(),
             blocks: GeneratedCppQObjectBlocks::default(),
+            has_qobject_macro: true,
         }
     }
 
@@ -170,7 +203,7 @@ mod tests {
             &[],
             "BaseClass".to_owned(),
             &["member1(1)".to_string(), "member2{ 2 }".to_string()],
-            &TypeNames::default(),
+            &type_names_with_qobject(),
         )
         .unwrap();
 
@@ -200,7 +233,7 @@ mod tests {
             &[],
             "BaseClass".to_owned(),
             &[],
-            &TypeNames::default(),
+            &type_names_with_qobject(),
         )
         .unwrap();
 
@@ -223,6 +256,37 @@ mod tests {
     }
 
     #[test]
+    fn default_constructor_no_qobject_macro() {
+        let mut qobject = qobject_for_testing();
+        qobject.has_qobject_macro = false;
+        let blocks = generate(
+            &qobject,
+            &[],
+            "BaseClass".to_owned(),
+            &[],
+            &type_names_with_qobject(),
+        )
+        .unwrap();
+
+        assert_empty_blocks(&blocks);
+        assert!(blocks.private_methods.is_empty());
+        assert_eq!(
+            blocks.methods,
+            vec![CppFragment::Pair {
+                header: "explicit MyObject();".to_string(),
+                source: formatdoc!(
+                    "
+                    MyObject::MyObject()
+                      : BaseClass()
+                      , ::rust::cxxqt1::CxxQtType<MyObjectRust>(::rust::createRs())
+                    {{ }}
+                    "
+                ),
+            }]
+        );
+    }
+
+    #[test]
     fn constructor_without_base_arguments() {
         let blocks = generate(
             &qobject_for_testing(),
@@ -232,7 +296,7 @@ mod tests {
             }],
             "BaseClass".to_owned(),
             &[],
-            &TypeNames::default(),
+            &type_names_with_qobject(),
         )
         .unwrap();
 
@@ -282,7 +346,7 @@ mod tests {
             }],
             "BaseClass".to_owned(),
             &["initializer".to_string()],
-            &TypeNames::default(),
+            &type_names_with_qobject(),
         )
         .unwrap();
 
@@ -336,7 +400,7 @@ mod tests {
             ],
             "BaseClass".to_owned(),
             &["initializer".to_string()],
-            &TypeNames::default(),
+            &type_names_with_qobject(),
         )
         .unwrap();