+++ crates/cxx-qt-lib-headers/include/gui/qfont.h
@@ -24,6 +24,8 @@ using QFontStyle = QFont::Style;
 using QFontHintingPreference = QFont::HintingPreference;
 using QFontCapitalization = QFont::Capitalization;
 using QFontSpacingType = QFont::SpacingType;
+using QFontStyleStrategy = QFont::StyleStrategy;
+using QFontStyleHint = QFont::StyleHint;
 
 } // namespace cxxqtlib1
 } // namespace rust