+++ crates/cxx-qt-lib/src/core/qstringlist.rs
@@ -98,9 +98,9 @@ pub struct QStringList {
     ///
     /// Qt5 QStringList has one pointer as a member
     /// Qt6 QStringList has one member, which contains two pointers and a size_t
-    #[cfg(qt_version_major = "5")]
+    #[cfg(cxxqt_qt_version_major = "5")]
     _space: MaybeUninit<usize>,
-    #[cfg(qt_version_major = "6")]
+    #[cfg(cxxqt_qt_version_major = "6")]
     _space: MaybeUninit<[usize; 3]>,
 }