+++ crates/cxx-qt-lib/src/gui/qfont.rs
@@ -73,6 +73,73 @@ mod ffi {
         AbsoluteSpacing,
     }
 
+    /// The style strategy tells the font matching algorithm what type of fonts should
+    /// be used to find an appropriate default family.
+    #[repr(i32)]
+    #[namespace = "rust::cxxqtlib1"]
+    #[derive(Debug)]
+    enum QFontStyleStrategy {
+        /// the default style strategy. It does not prefer any type of font
+        PreferDefault = 0x0001,
+        /// prefers bitmap fonts (as opposed to outline fonts).
+        PreferBitmap = 0x0002,
+        /// prefers device fonts.
+        PreferDevice = 0x0004,
+        /// prefers outline fonts (as opposed to bitmap fonts).
+        PreferOutline = 0x0008,
+        /// forces the use of outline fonts.
+        ForceOutline = 0x0010,
+        // don't antialias the fonts.
+        // PreferMatch = 0x0020,
+        // avoid subpixel antialiasing on the fonts if possible
+        // PreferQuality = 0x0040,
+        /// antialias if possible.
+        PreferAntialias = 0x0080,
+        /// don't antialias the fonts.
+        NoAntialias = 0x0100,
+        /// avoid subpixel antialiasing on the fonts if possible.
+        NoSubpixelAntialias = 0x0800,
+        /// Sometimes, a font will apply complex rules to a set of characters
+        /// in order to display them correctly. In some writing systems, such
+        /// as Brahmic scripts, this is required in order for the text to be legible,
+        /// but in e.g. Latin script, it is merely a cosmetic feature.
+        /// The PreferNoShaping flag will disable all such features when they are not
+        /// required, which will improve performance in most cases (since Qt 5.10).
+        PreferNoShaping = 0x1000,
+        /// If the font selected for a certain writing system does not contain a character
+        /// requested to draw, then Qt automatically chooses a similar looking font that
+        /// contains the character. The NoFontMerging flag disables this feature. Please note
+        /// that enabling this flag will not prevent Qt from automatically picking a
+        /// suitable font when the selected font does not support the writing system of the text.
+        NoFontMerging = 0x8000,
+    }
+
+    /// Style hints are used by the font matching algorithm to find an appropriate default
+    /// family if a selected font family is not available.
+    #[repr(i32)]
+    #[namespace = "rust::cxxqtlib1"]
+    #[derive(Debug)]
+    enum QFontStyleHint {
+        /// is a synonym for SansSerif.
+        Helvetica,
+        /// is a synonym for Serif.
+        Times,
+        /// a synonym for TypeWriter.
+        Courier,
+        /// the font matcher prefers decorative fonts.
+        OldEnglish,
+        /// the font matcher prefers system fonts.
+        System,
+        /// leaves the font matching algorithm to choose the family. This is the default.
+        AnyStyle,
+        /// the font matcher prefers fonts that map to the CSS generic font-family 'cursive'.
+        Cursive,
+        /// the font matcher prefers fonts that map to the CSS generic font-family 'monospace'.
+        Monospace,
+        /// the font matcher prefers fonts that map to the CSS generic font-family 'fantasy'.
+        Fantasy,
+    }
+
     unsafe extern "C++" {
         include!("cxx-qt-lib/qfont.h");
         type QFont = super::QFont;
@@ -176,6 +243,10 @@ mod ffi {
         #[rust_name = "set_fixed_pitch"]
         fn setFixedPitch(self: &mut QFont, enable: bool);
 
+        /// Sets the style strategy for the font to s.
+        #[rust_name = "set_style_strategy"]
+        fn setStyleStrategy(self: &mut QFont, strategy: QFontStyleStrategy);
+
         /// Set the preference for the hinting level of the glyphs to hintingPreference.
         #[rust_name = "set_hinting_preference"]
         fn setHintingPreference(self: &mut QFont, hintingPreference: QFontHintingPreference);
@@ -212,6 +283,11 @@ mod ffi {
         #[rust_name = "set_style"]
         fn setStyle(self: &mut QFont, style: QFontStyle);
 
+        /// Sets the style hint and strategy to hint and strategy, respectively.
+        /// Qt does not support style hints on X11 since this information is not provided by the window system.
+        #[rust_name = "set_style_hint"]
+        fn setStyleHint(self: &mut QFont, hint: QFontStyleHint, strategy: QFontStyleStrategy);
+
         /// Sets the style name of the font to styleName.
         #[rust_name = "set_style_name"]
         fn setStyleName(self: &mut QFont, styleName: &QString);
@@ -228,12 +304,21 @@ mod ffi {
         fn stretch(self: &QFont) -> i32;
 
         /// Returns true if strikeout has been set; otherwise returns false.
+        #[rust_name = "strike_out"]
         fn strikeOut(self: &QFont) -> bool;
+
+        /// Returns the StyleHint.
+        #[rust_name = "style_hint"]
+        fn styleHint(self: &QFont) -> QFontStyleHint;
 
         /// Returns the requested font style name. This can be used to match the font
         /// with irregular styles (that can't be normalized in other style properties).
         #[rust_name = "style_name"]
         fn styleName(self: &QFont) -> QString;
+
+        /// Returns the StyleStrategy.
+        #[rust_name = "style_strategy"]
+        fn styleStrategy(self: &QFont) -> QFontStyleStrategy;
 
         /// Returns true if underline has been set; otherwise returns false.
         fn underline(self: &QFont) -> bool;
@@ -250,6 +335,8 @@ mod ffi {
         type QFontHintingPreference;
         type QFontCapitalization;
         type QFontSpacingType;
+        type QFontStyleStrategy;
+        type QFontStyleHint;
 
         #[doc(hidden)]
         #[rust_name = "qfont_init_default"]
@@ -262,10 +349,22 @@ mod ffi {
         #[doc(hidden)]
         #[rust_name = "qfont_clone"]
         fn construct(font: &QFont) -> QFont;
+
+        #[doc(hidden)]
+        #[rust_name = "qfont_eq"]
+        fn operatorEq(a: &QFont, b: &QFont) -> bool;
+
+        #[doc(hidden)]
+        #[rust_name = "qfont_to_qstring"]
+        fn toQString(value: &QFont) -> QString;
+
     }
 }
 
-pub use ffi::{QFontCapitalization, QFontHintingPreference, QFontSpacingType, QFontStyle};
+pub use ffi::{
+    QFontCapitalization, QFontHintingPreference, QFontSpacingType, QFontStyle, QFontStyleHint,
+    QFontStyleStrategy,
+};
 
 #[repr(C)]
 pub struct QFont {
@@ -290,6 +389,20 @@ impl Clone for QFont {
         ffi::qfont_clone(self)
     }
 }
+
+impl std::fmt::Display for QFont {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", ffi::qfont_to_qstring(self))
+    }
+}
+
+impl PartialEq for QFont {
+    fn eq(&self, other: &Self) -> bool {
+        ffi::qfont_eq(self, other)
+    }
+}
+
+impl Eq for QFont {}
 
 impl QFont {
     /// Returns the bounding rectangle of the current clip if there is a clip;