+++ crates/cxx-qt-lib/src/gui/qimage.rs
@@ -237,6 +237,10 @@ mod ffi {
         #[doc(hidden)]
         #[rust_name = "qimage_cache_key"]
         fn qimageCacheKey(image: &QImage) -> i64;
+
+        #[doc(hidden)]
+        #[rust_name = "qimage_eq"]
+        fn operatorEq(a: &QImage, b: &QImage) -> bool;
     }
 }
 
@@ -250,9 +254,9 @@ pub use ffi::{QImageFormat, QImageInvertMode};
 pub struct QImage {
     // Static checks on the C++ side ensure this is true.
     // See qcolor.cpp
-    #[cfg(qt_version_major = "5")]
+    #[cfg(cxxqt_qt_version_major = "5")]
     _data: MaybeUninit<[usize; 4]>,
-    #[cfg(qt_version_major = "6")]
+    #[cfg(cxxqt_qt_version_major = "6")]
     _data: MaybeUninit<[usize; 3]>,
 }
 
@@ -269,6 +273,14 @@ impl Default for QImage {
         ffi::qimage_init_default()
     }
 }
+
+impl PartialEq for QImage {
+    fn eq(&self, other: &Self) -> bool {
+        ffi::qimage_eq(self, other)
+    }
+}
+
+impl Eq for QImage {}
 
 // Safety:
 //