+++ crates/cxx-qt-lib/src/gui/qpainterpath.rs
@@ -4,6 +4,7 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 use crate::QPointF;
 use cxx::{type_id, ExternType};
+use std::fmt;
 use std::mem::MaybeUninit;
 
 #[cxx::bridge]
@@ -188,6 +189,14 @@ mod ffi {
         #[doc(hidden)]
         #[rust_name = "qpainterpath_drop"]
         fn drop(painterPath: &mut QPainterPath);
+
+        #[doc(hidden)]
+        #[rust_name = "qpainterpath_eq"]
+        fn operatorEq(a: &QPainterPath, b: &QPainterPath) -> bool;
+
+        #[doc(hidden)]
+        #[rust_name = "qpainterpath_to_qstring"]
+        fn toQString(value: &QPainterPath) -> QString;
     }
 }
 
@@ -221,6 +230,20 @@ impl From<&QPointF> for QPainterPath {
         ffi::qpainterpath_from_qpointf(p)
     }
 }
+
+impl PartialEq for QPainterPath {
+    fn eq(&self, other: &Self) -> bool {
+        ffi::qpainterpath_eq(self, other)
+    }
+}
+
+impl fmt::Display for QPainterPath {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", ffi::qpainterpath_to_qstring(self))
+    }
+}
+
+impl Eq for QPainterPath {}
 
 // Safety:
 //