+++ crates/cxx-qt-lib/src/gui/qpen.rs
@@ -3,6 +3,7 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 use cxx::{type_id, ExternType};
+use std::fmt;
 use std::mem::MaybeUninit;
 
 #[cxx::bridge]
@@ -22,6 +23,8 @@ mod ffi {
         type QColor = crate::QColor;
         include!("cxx-qt-lib/qbrush.h");
         type QBrush = crate::QBrush;
+        include!("cxx-qt-lib/qstring.h");
+        type QString = crate::QString;
 
         /// Returns the brush used to fill strokes generated with this pen.
         fn brush(self: &QPen) -> QBrush;
@@ -122,14 +125,22 @@ mod ffi {
         #[doc(hidden)]
         #[rust_name = "qpen_clone"]
         fn construct(pen: &QPen) -> QPen;
+
+        #[doc(hidden)]
+        #[rust_name = "qpen_eq"]
+        fn operatorEq(a: &QPen, b: &QPen) -> bool;
+
+        #[doc(hidden)]
+        #[rust_name = "qpen_to_qstring"]
+        fn toQString(value: &QPen) -> QString;
     }
 }
 
 #[repr(C)]
 pub struct QPen {
-    #[cfg(qt_version_major = "5")]
+    #[cfg(cxxqt_qt_version_major = "5")]
     _cspec: MaybeUninit<[i32; 2]>,
-    #[cfg(qt_version_major = "6")]
+    #[cfg(cxxqt_qt_version_major = "6")]
     _cspec: MaybeUninit<i32>,
 }
 
@@ -152,6 +163,14 @@ impl Clone for QPen {
     }
 }
 
+impl PartialEq for QPen {
+    fn eq(&self, other: &Self) -> bool {
+        ffi::qpen_eq(self, other)
+    }
+}
+
+impl Eq for QPen {}
+
 impl From<&ffi::QColor> for QPen {
     fn from(color: &ffi::QColor) -> Self {
         ffi::qpen_init_from_qcolor(color)
@@ -161,6 +180,12 @@ impl From<&ffi::QColor> for QPen {
 impl From<&ffi::PenStyle> for QPen {
     fn from(penstyle: &ffi::PenStyle) -> Self {
         ffi::qpen_init_from_penstyle(penstyle)
+    }
+}
+
+impl fmt::Display for QPen {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", ffi::qpen_to_qstring(self))
     }
 }