+++ crates/cxx-qt/build.rs
@@ -52,13 +52,11 @@ fn main() {
 
     // Note, ensure our settings stay in sync across cxx-qt, cxx-qt-build, and cxx-qt-lib
     builder.cpp(true);
+    builder.std("c++17");
     // MSVC
-    builder.flag_if_supported("/std:c++17");
     builder.flag_if_supported("/Zc:__cplusplus");
     builder.flag_if_supported("/permissive-");
     builder.flag_if_supported("/bigobj");
-    // GCC + Clang
-    builder.flag_if_supported("-std=c++17");
     // MinGW requires big-obj otherwise debug builds fail
     builder.flag_if_supported("-Wa,-mbig-obj");