+++ tests/qt_types_standalone/cpp/main.cpp
@@ -26,6 +26,7 @@
 #include "qpersistentmodelindex.h"
 #include "qpoint.h"
 #include "qpointf.h"
+#include "qpolygon.h"
 #include "qqmlapplicationengine.h"
 #include "qqmlengine.h"
 #include "qrect.h"
@@ -92,6 +93,7 @@ main(int argc, char* argv[])
   runTest(QScopedPointer<QObject>(new QVector2DTest));
   runTest(QScopedPointer<QObject>(new QVector3DTest));
   runTest(QScopedPointer<QObject>(new QVector4DTest));
+  runTest(QScopedPointer<QObject>(new QPolygonTest));
 
   return status;
 }